ai_backend/requirements.txt
fastapi 
uvicorn 
tensorflow 
pillow 
numpy 
python-dotenv 
supabase 
python-multipart
